Ascii writing works.
Ascii reading works one column at a time.

Discovery:  With ascii you can write numbers with different precisions,
but cfitsio thinks the columns are always i4 or f8, period.  That sucks.
So in order to read as i8 we would have to force the data type some how.
Maybe examine the tform for ascii and determine the precision from
that instead of the usual means.