Fix more 32-bit sizeof issues On Mips and PowerPC, sizeof(long) == 4, and sizeof(npy_intp) == 4, so that one cannot replace `npy_int64` with `long` there.
On Mips, this leads to a Floating point exception in `testImageSlice`
(division by 0, since the stepsize is then errornously set to zero).

On PowerPC, this is catched, but leads to a failure `testImageSlice`. On both
platforms, `testImageWriteReadFromDimsChunks` and `testImageWriteReadFromDims`
fail as well.

This patch uses `npy_int64` on the places where these problems appear.