starting work on aligned read
The idea here is to create the output rec array and get
references to the columns and send that list to the reader

This way it is unimportant if the array was declared to
be aligned or not.  Internally the array access routines
are used to guarantee the right offsets are used.