Support fits_open_memfile/fits_create_memfile
It is occasionally useful to be able to use fitsio on files which you
have already read into memory through other means. The cfitsio library
supports this with fits_open_memfile/fits_create_memfile. Support is now
present for this method of operation.

This feature was developed with the use case of a thumbnail server in
mind. In our implementation, FITS files are received over HTTPS from the
network, immediately converted into a JPEG, and then sent out over HTTPS
to the client. By keeping everything in memory, the requests are
processed a little bit faster, and we don't need to worry about filling
up our disks with temporary files over time.

The implementation is somewhat suboptimal, because I could not find a
way in the Python interface that lets me use realloc() on a bytearray
(or similar Python object). Therefore, I had to create a copy of the
bytearray and use it instead, only copying back into the bytearray when
calling the close() method.

If we change this feature so that the only supported mode is read-only,
then this suboptimal behavior can go away (it becomes zero copy). Zero
copy mode breaks the read_raw() accessor (SEGV in memcpy()). If I remove
zero copy mode and always copy, then everything I've tested works fine.

However, as an initial draft, I thought it would be beneficial to show
that we can do both read-only and read-write modes.

Signed-off-by: Ira W. Snyder <isnyder@lco.global>